Update formatting and page organization
This change updates the organization of the page, specifically the listing
of space access protocols, to match the alphabetized drop-down menu in the
Storage Service itself. I've also reformatted the field lists to match similar
lists in other parts of the Archivematica documentation. Some minor changes
were made to the contents of the page, but only for clarity.