Formatting and clarification updates (#17)
Most of this change is formatting and cleanup, but there is one somewhat major
change: I've removed the page users.rst
(see https://www.archivematica.org/en/docs/storage-service-0.13/users/#users)
because I didn't think it was very helpful! I've removed the page from indices
and toctrees and changed the name of the Administrators page to remove the
idea of an Administrator/User dichotomy, as we have in archivematica-docs.