Update Administration tab section
This change updates the Administration tab section of the Storage
Service documentation, adding in some info that was not previously
documented - namely the object counting button and the recovery
notification fields.

I've also done some basic clean up in other areas of the page, for
formatting and consistency.

Connected to archivematica/Issues#657