Update index page with headings and links
This change reformats the main page of the Storage Service documentation
to add headings for each of the main sections as well as links to
subsections, for easier navigation.

I've made some small changes on sub-pages so that I could add links to
subsections, but further updates will come in future PRs.